Revert delete settingscontroller
