+++ src/java/SongController.java
@@ -4,7 +4,6 @@ import javafx.scene.Node;
 import javafx.scene.control.Button;
 import javafx.scene.control.TextField;
 import javafx.event.ActionEvent;
-import javafx.stage.DirectoryChooser;
 import javafx.stage.FileChooser;
 import javafx.stage.Stage;
 import model.LocalFilesModel;
@@ -22,13 +21,7 @@ public class SongController implements Initializable {
 
     @FXML public Button btnSelectFile;
 
-    @FXML public TextField txtFieldDirectory;
-
-    @FXML public Button btnSelectDirectory;
-    @FXML public Button btnApply;
-
     String songPath = null;
-    Path localDirPath = null;
     LocalFilesModel localFilesModel;
 
     public SongController()
@@ -68,38 +61,6 @@ public class SongController implements Initializable {
     }
 
     public void onEditSong(ActionEvent event) {
-
-        ((Node)(event.getSource())).getScene().getWindow().hide();
-    }
-
-    /**
-     * Opens a DirectoryChooser, and sets the accompanying TextField's text as the directory.
-     * Should the user choose to manipulate the chosen directory, it parses the text in the TextField
-     * as a path, which is then used by other classes.
-     */
-    @FXML
-    public void onSelectDirectory(ActionEvent event){
-        DirectoryChooser dc = new DirectoryChooser();
-
-        File selectedDirectory = dc.showDialog(new Stage());
-
-        if (selectedDirectory == null)
-            return;
-
-        txtFieldDirectory.setText(selectedDirectory.getAbsolutePath());
-
-        Path path = Path.of(txtFieldDirectory.getText());
-    }
-
-    @FXML
-    public void onApplySettings(ActionEvent event){
-        if(!txtFieldDirectory.getText().isEmpty() &&
-                !txtFieldDirectory.getText().equals(null))
-        {
-            localFilesModel.getAllLocalSongs(Path.of(txtFieldDirectory.getText()));
-            localDirPath = Path.of(txtFieldDirectory.getText());
-            localFilesModel.setLocalDirPathInModel(localDirPath);
-        }
 
         ((Node)(event.getSource())).getScene().getWindow().hide();
     }
